Final_Project/requirements.txt
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
Pillow>=8.0.0
gradio>=3.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
tqdm>=4.45.0
gdown>=4.4.0  # For downloading the dataset and models